Damage System Improvements
- Damage is now a float variable
- Added Damage Resistance to hosts
- Added Adjusted damage to enemies based on base host damage
- Added Spread shot parasite

+++ Brackieys Jam/Assets/Code/EnemyControllers/BaseEnemyController.cs
@@ -11,7 +11,7 @@ public class BaseEnemyController : MonoBehaviour
     [SerializeField] protected Rigidbody2D MyRigidBody;
     [Space]
     [SerializeField] protected EnemyState State = EnemyState.Idle;
-    [SerializeField] protected int Health = 5;
+    [SerializeField] protected float Health = 5;
     [SerializeField] protected float MovementSpeed = 3;
     [SerializeField] protected float StateDuration;
